Refactor: Update shopping list app bar structure and introduce new components
