+++ test/features/shoppinglist/presentation/shopping_list_page_test.dart
@@ -6,7 +6,8 @@ import 'package:fake_cloud_firestore/fake_cloud_firestore.dart';
 import 'package:mealtrack/core/presentation/widgets/feature_sliver_app_bar.dart';
 import 'package:mealtrack/features/shoppinglist/domain/shopping_list_item.dart';
 import 'package:mealtrack/features/shoppinglist/presentation/shopping_list_page.dart';
-import 'package:mealtrack/features/shoppinglist/presentation/widgets/shopping_list_sliver_app_bar.dart';
+import 'package:mealtrack/features/shoppinglist/presentation/widgets/shopping_list_appbar/shopping_list_cost_summary.dart';
+import 'package:mealtrack/features/shoppinglist/presentation/widgets/shopping_list_appbar/shopping_list_sliver_app_bar.dart';
 import 'package:mealtrack/features/shoppinglist/data/shopping_list_repository.dart';
 import 'package:mealtrack/features/shoppinglist/data/category_stats_repository.dart';
 import 'package:mealtrack/features/shoppinglist/domain/category_suggestion.dart';
